Move the interface in Service Interface Folder and add the response entity in /getAllBatchesData API
