+++ src/main/java/com/japanese/appliaction/service/serviceInterface/TeacherService.java
@@ -1,4 +1,4 @@
-package com.japanese.appliaction.service;
+package com.japanese.appliaction.service.serviceInterface;
 
 import java.util.List;
